+++ functions/lib/index.js
@@ -60,29 +60,24 @@ exports.notify = (0, https_1.onRequest)(
     try {
         (0, messaging_1.getMessaging)()
             .sendEachForMulticast(message)
-            .then((response) => {
-            const failedTokens = [];
-            const successTokens = [];
+            .then(async (response) => {
+            const promises = [];
             response.responses.forEach((resp, idx) => {
                 if (resp && idx < registrationTokens.length) {
                     if (!resp.success) {
-                        failedTokens.push(registrationTokens[idx]);
+                        promises.push(tokenDispacher(registrationTokens[idx], false, 'n/a'));
                     }
                     else {
-                        successTokens.push(registrationTokens[idx]);
+                        promises.push(tokenDispacher(registrationTokens[idx], true, text));
                     }
                 }
             });
-            if (failedTokens.length > 0) {
-                failedTokens.forEach(async (token) => await removeToken(token));
-            }
-            if (successTokens.length > 0) {
-                successTokens.forEach(async (token) => await messageSent(token, text));
-            }
-            else if (successTokens.length === 0) {
+            if (promises.length === 0) {
                 res.status(200).send('No active subscribers found');
                 logger.info(`No active subscribers found. No message sent`);
+                return;
             }
+            await Promise.all(promises);
         });
     }
     catch (error) {
@@ -90,37 +85,29 @@ exports.notify = (0, https_1.onRequest)(
         res.status(500).json({ error: error.message });
     }
 });
-const removeToken = async (token) => {
+const tokenDispacher = async (token, status, msg) => {
     if (token === undefined)
         return;
     const docRef = (0, firestore_1.getFirestore)().collection('Device').doc(token);
     const doc = await docRef.get();
     const data = doc.data();
-    const diff = Date.now() - data?.timestamp;
-    logger.info(`Removed token for ${data?.email} age ` + Math.floor(diff / 86400000));
-    await docRef.delete();
+    let text = 'successfully sent';
+    if (status) {
+        logger.info(`Message sent to ${data?.email}`);
+    }
+    else {
+        const diff = Date.now() - data?.timestamp.toMillis();
+        text = 'removed token age ' + Math.floor(diff / 86400000);
+        logger.info(`Removed token for ${data?.email} age ` + Math.floor(diff / 86400000));
+        await docRef.delete();
+    }
     await (0, firestore_1.getFirestore)()
         .collection('Message')
         .add({
         email: data?.email,
-        text: '-',
-        status: 'removed token age ' + Math.floor(diff / 86400000),
-        timestamp: firestore_2.Timestamp.fromDate(new Date()),
-    });
-};
-const messageSent = async (token, text) => {
-    if (token === undefined)
-        return;
-    const docRef = (0, firestore_1.getFirestore)().collection('Device').doc(token);
-    const doc = await docRef.get();
-    const data = doc.data();
-    logger.info(`Message sent to ${data?.email}`);
-    await (0, firestore_1.getFirestore)()
-        .collection('Message')
-        .add({
-        email: data?.email,
+        message: msg,
+        status: status,
         text: text,
-        status: 'successfully sent',
         timestamp: firestore_2.Timestamp.fromDate(new Date()),
     });
 };
